Split address classifier from core util
