+++ ydb/core/util/address_classifier/ya.make
@@ -1,0 +1,13 @@
+LIBRARY()
+
+SRCS(
+    address_classifier.cpp
+    address_classifier.h
+)
+
+PEERDIR(
+    library/cpp/ipmath
+    ydb/core/protos
+)
+
+END()
